requirements.txt
sionna
tensorflow-cpu==2.15.1
torch
torchvision
torchaudio
ipykernel
jupyterlab
lightly
